@@ -1,23 +1,39 @@
 {
-    "name": "1.-node",
+    "name": "node.js_mvc_typescript",
     "version": "1.0.0",
-    "description": "Node.JS Tutorial",
+    "description": "Node.js MVC made with TypeScript & EJS",
     "main": "app.js",
     "scripts": {
         "test": "mocha",
-        "start": "node app.js",
-        "start:dev": "nodemon app.js"
+        "start": "node ./build/app.js",
+        "start:dev": "ts-node-dev ./src/app.ts"
     },
+    "keywords": [
+        "Node.js",
+        "MVC",
+        "TypeScript"
+    ],
     "author": "Ion Munteanu",
     "license": "ISC",
     "devDependencies": {
-        "chai": "^4.3.8",
+        "@types/compression": "^1.7.5",
+        "@types/connect-flash": "^0.0.40",
+        "@types/connect-mongodb-session": "^2.4.7",
+        "@types/csurf": "^1.11.5",
+        "@types/express": "^4.17.21",
+        "@types/express-session": "^1.17.10",
+        "@types/morgan": "^1.9.9",
+        "@types/multer": "^1.4.11",
+        "@types/node": "^20.10.6",
+        "@types/pdfkit": "^0.13.3",
+        "chai": "^5.0.0",
         "ejs": "^3.1.9",
         "mocha": "^10.2.0",
-        "nodemon": "^3.0.1",
-        "typescript": "^5.1.6"
+        "ts-node-dev": "^2.0.0",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
+        "@types/mongoose": "^5.11.97",
         "bcryptjs": "^2.4.3",
         "compression": "^1.7.4",
         "connect-flash": "^0.1.1",
@@ -26,15 +42,15 @@
         "express": "^4.18.2",
         "express-session": "^1.17.3",
         "express-validator": "^7.0.1",
-        "helmet": "^7.0.0",
-        "mongodb": "^5.8.0",
-        "mongoose": "^7.4.4",
+        "helmet": "^7.1.0",
+        "mongodb": "^6.3.0",
+        "mongoose": "^8.0.3",
         "morgan": "^1.10.0",
-        "multer": "^1.4.5-lts.1",
-        "mysql2": "^3.6.0",
-        "nodemailer": "^6.9.4",
-        "pdfkit": "^0.13.0",
-        "sequelize": "^6.32.1",
-        "stripe": "^8.222.0"
+        "multer": "1.4.5-lts.1",
+        "mysql2": "^3.6.5",
+        "nodemailer": "^6.9.8",
+        "pdfkit": "^0.14.0",
+        "sequelize": "^6.35.2",
+        "stripe": "^14.11.0"
     }
 }
