@@ -1,6 +1,6 @@
-const mongoose = require("mongoose");
+import mongoose from 'mongoose'
 
-const Schema = mongoose.Schema;
+const Schema = mongoose.Schema
 
 const userSchema = new Schema({
     email: {
@@ -18,49 +18,50 @@ const userSchema = new Schema({
             {
                 productId: {
                     type: Schema.Types.ObjectId,
-                    ref: "Product",
+                    ref: 'Product',
                     required: true,
                 },
                 quantity: { type: Number, required: true },
             },
         ],
     },
-});
+})
 
 userSchema.methods.addToCart = function (product) {
     const cartProductIndex = this.cart.items.findIndex((cp) => {
-        return cp.productId.toString() === product._id.toString();
-    });
-    let newQuantity = 1;
-    const updatedCartItems = [...this.cart.items];
+        return cp.productId.toString() === product._id.toString()
+    })
+    let newQuantity = 1
+    const updatedCartItems = [...this.cart.items]
 
     if (cartProductIndex >= 0) {
-        newQuantity = this.cart.items[cartProductIndex].quantity + 1;
-        updatedCartItems[cartProductIndex].quantity = newQuantity;
+        newQuantity = this.cart.items[cartProductIndex].quantity + 1
+        updatedCartItems[cartProductIndex].quantity = newQuantity
     } else {
         updatedCartItems.push({
             productId: product._id,
             quantity: newQuantity,
-        });
+        })
     }
     const updatedCart = {
         items: updatedCartItems,
-    };
-    this.cart = updatedCart;
-    return this.save();
-};
+    }
+    this.cart = updatedCart
+    return this.save()
+}
 
 userSchema.methods.removeFromCart = function (productId) {
     const updatedCartItems = this.cart.items.filter((item) => {
-        return item.productId.toString() !== productId.toString();
-    });
-    this.cart.items = updatedCartItems;
-    return this.save();
-};
+        return item.productId.toString() !== productId.toString()
+    })
+    this.cart.items = updatedCartItems
+    return this.save()
+}
 
 userSchema.methods.clearCart = function () {
-    this.cart = { items: [] };
-    return this.save();
-};
+    this.cart = { items: [] }
+    return this.save()
+}
 
-module.exports = mongoose.model("User", userSchema);
+// module.exports = mongoose.model('User', userSchema)
+export default mongoose.model('User', userSchema)
